src/requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
drf-nested-routers==0.94.2
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
